+++ .gitignore
@@ -11,3 +11,5 @@ __pycache__/
 
 # VSCode
 .vscode
+
+notes/*
