+++ AUTHORS.rst
@@ -1,0 +1,6 @@
+Authors
+=======
+
+* Zixuan Zhong         <https://github.com/ZixuanZhong>
+* Justin Presley       <https://github.com/justincpresley>
+* Daniel Achee         <https://github.com/daniel-achee>
