+++ LICENSE.rst
