--- README.md
@@ -1,20 +1,0 @@
-# ndn-distributed-repo
-
-## Dependencies
-
-```
-pip install pytest
-pip install python-ndn ndn-svs ndn-python-repo
-```
-
-Go to [python-ndn repo](https://github.com/zjkmxy/python-ndn) if you have issues installing
-
-## To Run Tests
-
-```
-$pytest
-```
-
-## Troubleshooting
-
-Ensure ndn_distributed_repo is on PYTHON_PATH
