+++ README.rst
@@ -1,0 +1,10 @@
+hydra: a NDN distributed repo
+=============================
+
+A Named Data Networking (NDN) Distributed Repo implementation using python-ndn_.
+
+Please see our documentation_.
+
+.. _python-ndn: https://github.com/named-data/python-ndn
+
+.. _documentation: https://hydra.readthedocs.io/en/latest
