--- client/example/t.txt
@@ -1,1 +1,0 @@
-abcdefghijklmnopqrstuvwxyz
