+++ client/functions/query.py
@@ -13,7 +13,7 @@ import os
 from ndn.app import NDNApp
 from ndn.encoding import FormalName, Component, Name, ContentType
 from ndn.types import InterestNack, InterestTimeout, InterestCanceled, ValidationFailure
-from ndn_distributed_repo.protocol import File, FileList
+from hydra.protocol import File, FileList
 
 class QueryClient(object):
     def __init__(self, app: NDNApp, client_prefix: FormalName, repo_prefix: FormalName) -> None:
