+++ docs/index.rst
@@ -1,0 +1,31 @@
+.. Hydra - An NDN Distributed Repo documentation master file, created by
+   sphinx-quickstart on Thu Jun 24 02:19:18 2021.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to hydra_'s documentation!
+===========================================================
+
+Hydra is a distributed repo implementation in NDN.
+Our design focuses on three attributes: resiliency, scalability,
+usability, and efficiency.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   src/readme.rst
+   src/install.rst
+   src/configuration.rst
+   src/examples.rst
+   src/authors.rst
+   src/future.rst
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+
+.. _hydra: https://github.com/UCLA-IRL/hydra
