+++ docs/requirements.txt
@@ -1,0 +1,6 @@
+sphinx-autodoc-typehints>=1.12.0
+sphinx-rtd-theme>=0.5.2
+python-ndn>=0.3a1.post1
+PyYAML>=5.4.1
+ndn-svs >= 0.2.1
+ndn-python-repo >= 0.2a5
