+++ docs/src/authors.rst
@@ -1,0 +1,1 @@
+.. include:: ../../AUTHORS.rst
