+++ docs/src/configuration.rst
@@ -1,0 +1,14 @@
+Configuration
+=============
+
+Repo Prefix Strategy
+--------------------
+
+Before running hydra on any node, we must register the repo prefix as multicast.
+You can do that by doing the following..
+
+.. code-block:: bash
+
+    $ nfdc strategy set /<repo-prefix>/group /localhost/nfd/strategy/multicast/%FD%03
+
+More info on different strategies found `here <https://named-data.net/doc/NFD/current/manpages/nfdc-strategy.html>`_.
