+++ docs/src/future.rst
@@ -1,0 +1,17 @@
+Future Plans
+============
+
+Implemention Advancements
+-------------------------
+
+There are many improvements to hydra that need to be made.
+    * Improve File Fetching
+    * Implement Security
+    * More Tests
+    * Better documentation
+
+Performance Testing
+-------------------
+
+Hydra is new in concept and implementation, so testing it and finding performance
+results in various environments is critical for the continuation of this project.
