+++ docs/src/install.rst
@@ -1,0 +1,53 @@
+Install
+=======
+
+Dependencies
+------------
+
+* Install / Setup NFD_.
+
+Instruction for use
+--------------------
+
+(WIP) Install the latest release with pip:
+
+.. code-block:: bash
+
+    $ /usr/bin/pip3 install ndn-hydra
+
+Optionally, you can install the latest development version from source:
+(note: also add root folder of this github repo to PYTHONPATH environment variable)
+
+.. code-block:: bash
+
+    $ git clone https://github.com/UCLA-IRL/hydra.git
+    $ cd hydra && /usr/bin/pip3 install -e .
+
+
+Instruction for developers
+--------------------------
+
+Setup virtual environment with editable installation:
+
+.. code-block:: bash
+
+    $ python3 -m venv venv
+    $ . venv/bin/activate
+    $ pip3 install -e .
+
+Run all tests:
+
+.. code-block:: bash
+
+    $ pip3 install pytest
+    $ pytest
+
+Compile the documentation with Sphinx:
+
+.. code-block:: bash
+
+    $ cd docs && pip3 install -r requirements.txt
+    $ make html
+    $ open _build/html/index.html
+
+.. _NFD: https://named-data.net/doc/NFD/current/INSTALL.html
