+++ docs/src/readme.rst
@@ -1,0 +1,1 @@
+.. include:: ../../README.rst
