+++ hydra/handle_protocol/delete_command_handle.py
