+++ hydra/repo_messages/add.py
@@ -1,5 +1,5 @@
-from ndn_distributed_repo.data_storage.data_storage import DataStorage
-from ndn_distributed_repo.global_view.global_view import GlobalView
+from hydra.data_storage.data_storage import DataStorage
+from hydra.global_view.global_view import GlobalView
 from ndn.encoding import *
 from .message_base import MessageBodyBase
 from .store import StoreMessageBodyTlv
